色々リファクタリング
