+++ dispatch.py
@@ -1,0 +1,3 @@
+from src.deadline_reporting import deadline_reporting
+
+deadline_reporting()
