+++ requirements.txt
@@ -3,6 +3,7 @@ certifi==2024.2.2
 charset-normalizer==3.3.2
 click==8.1.7
 idna==3.6
+iniconfig==2.0.0
 Jinja2==3.1.3
 loguru==0.7.2
 MarkupSafe==2.1.5
@@ -10,6 +11,8 @@ mypy-extensions==1.0.0
 packaging==23.2
 pathspec==0.12.1
 platformdirs==4.2.0
+pluggy==1.4.0
+pytest==8.0.0
 requests==2.31.0
 slack_sdk==3.26.2
 urllib3==2.2.0
